Revert "changed codebase from long to long long"
This reverts commit 711720d58354ed210253b013d95167c1db1edf22.

+++ includes/mmap_graph_loader.h
@@ -15,14 +15,14 @@ class MMapGraphLoader
 public:
     struct Edge
     {
-        long long from;
-        long long to;
+        long from;
+        long to;
         double weight;
     };
     
     static int LoadEdgelistMMap(const std::string &filename,
                                 igraph_t *graph,
-                                std::unordered_map<long long,long long>* node_id_to_new_node_id_map,
+                                std::unordered_map<long,long>* node_id_to_new_node_id_map,
                                 bool has_weights)
     {
         int fd = open(filename.c_str(), O_RDONLY);
@@ -62,7 +62,7 @@ public:
         std::vector<Edge> edges;
         edges.reserve(estimated_edges);
 
-        long long next_node_id = 0;
+        long next_node_id = 0;
 
         while (ptr < end)
         {
@@ -71,12 +71,12 @@ public:
             if (ptr >= end)
                 break;
 
-            long long from = ParseLong(ptr, end);
+            long from = ParseLong(ptr, end);
 
             while (ptr < end && (*ptr == '\t' || *ptr == ' '))
                 ptr++;
 
-            long long to = ParseLong(ptr, end);
+            long to = ParseLong(ptr, end);
 
             double weight = 1.0;
 
@@ -116,8 +116,8 @@ public:
     }
 
     static int LoadClusteringMMap(const std::string &filename,
-                                  std::unordered_map<long long, long long>* node_to_cluster,
-                                  std::unordered_map<long long, long long> &node_id_to_new_node_id_cluster)
+                                  std::unordered_map<long, long>* node_to_cluster,
+                                  std::unordered_map<long, long> &node_id_to_new_node_id_cluster)
     {
         int fd = open(filename.c_str(), O_RDONLY);
         if (fd == -1) {
@@ -147,9 +147,9 @@ public:
             SkipToNextLine(ptr, end);
         }
         size_t estimated_clusters = EstimateEdgeCount(mapped, file_size);
-        std::unordered_map<long long, long long> cluster_id_to_new_cluster_id_map;
+        std::unordered_map<long, long> cluster_id_to_new_cluster_id_map;
 
-        long long cluster_id_new = 0;
+        long cluster_id_new = 0;
         while (ptr < end) 
         {
             while (ptr < end && (*ptr == '\n' || *ptr == '\r'))
@@ -157,12 +157,12 @@ public:
             if (ptr >= end)
                 break;
 
-            long long node = ParseLong(ptr, end);
+            long node = ParseLong(ptr, end);
 
             while (ptr < end && (*ptr == '\t' || *ptr == ' '))
                 ptr++;
 
-            long long cluster = ParseLong(ptr, end);
+            long cluster = ParseLong(ptr, end);
 
             if (cluster_id_to_new_cluster_id_map.find(cluster) == cluster_id_to_new_cluster_id_map.end()) {
                 cluster_id_to_new_cluster_id_map[cluster] = cluster_id_new++;
@@ -179,9 +179,9 @@ public:
     }
 
 private:
-    static inline long long ParseLong(const char *&ptr, const char *end)
+    static inline long ParseLong(const char *&ptr, const char *end)
     {
-        long long result = 0;
+        long result = 0;
         while (ptr < end && (*ptr == ' ' || *ptr == '\t'))
             ptr++;
